backend/requirements.txt
upstash-redis
python-decouple
jupyter
replicate
pillow
fastapi[standard]
fastapi-limiter
# support .heic images -> see readme
# pillow-heif 